requirements.txt
cbor2
dnspython[idna]
